+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\index.js":"1","C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\reportWebVitals.js":"2","C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\App.js":"3","C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\components\\CityCard.js":"4","C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\components\\Navbar.js":"5","C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\components\\ImageAndWelcome.js":"6","C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\components\\SearchCity.js":"7","C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\components\\CityList.js":"8","C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\components\\Footer\\Footer.js":"9"},{"size":500,"mtime":499162500000,"results":"10","hashOfConfig":"11"},{"size":362,"mtime":499162500000,"results":"12","hashOfConfig":"11"},{"size":1366,"mtime":1608311112552,"results":"13","hashOfConfig":"11"},{"size":465,"mtime":1608307842379,"results":"14","hashOfConfig":"11"},{"size":235,"mtime":1608308022035,"results":"15","hashOfConfig":"11"},{"size":1096,"mtime":1608308278091,"results":"16","hashOfConfig":"11"},{"size":902,"mtime":1608310396206,"results":"17","hashOfConfig":"11"},{"size":448,"mtime":1608310608163,"results":"18","hashOfConfig":"11"},{"size":1555,"mtime":1608311302970,"results":"19","hashOfConfig":"11"},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"tawo2g",{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\index.js",[],"C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\reportWebVitals.js",[],"C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\App.js",[],"C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\components\\CityCard.js",[],["39","40"],"C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\components\\Navbar.js",[],"C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\components\\ImageAndWelcome.js",[],"C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\components\\SearchCity.js",[],"C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\components\\CityList.js",[],"C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\components\\Footer\\Footer.js",[],{"ruleId":"41","replacedBy":"42"},{"ruleId":"43","replacedBy":"44"},"no-native-reassign",["45"],"no-negated-in-lhs",["46"],"no-global-assign","no-unsafe-negation"]
+[{"C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\index.js":"1","C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\reportWebVitals.js":"2","C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\App.js":"3","C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\components\\CityCard.js":"4","C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\components\\Navbar.js":"5","C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\components\\ImageAndWelcome.js":"6","C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\components\\SearchCity.js":"7","C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\components\\CityList.js":"8","C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\components\\Footer\\Footer.js":"9","C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\pages\\Home.js":"10"},{"size":500,"mtime":499162500000,"results":"11","hashOfConfig":"12"},{"size":362,"mtime":499162500000,"results":"13","hashOfConfig":"12"},{"size":332,"mtime":1608311827293,"results":"14","hashOfConfig":"12"},{"size":465,"mtime":1608307842379,"results":"15","hashOfConfig":"12"},{"size":235,"mtime":1608308022035,"results":"16","hashOfConfig":"12"},{"size":1096,"mtime":1608308278091,"results":"17","hashOfConfig":"12"},{"size":902,"mtime":1608310396206,"results":"18","hashOfConfig":"12"},{"size":448,"mtime":1608310608163,"results":"19","hashOfConfig":"12"},{"size":1555,"mtime":1608311302970,"results":"20","hashOfConfig":"12"},{"size":1214,"mtime":1608311790473,"results":"21","hashOfConfig":"12"},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"tawo2g",{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"30"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"30"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"30"},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"39","messages":"40","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"41","messages":"42","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\index.js",[],"C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\reportWebVitals.js",[],"C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\App.js",[],"C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\components\\CityCard.js",[],["43","44"],"C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\components\\Navbar.js",[],"C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\components\\ImageAndWelcome.js",[],"C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\components\\SearchCity.js",[],"C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\components\\CityList.js",[],"C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\components\\Footer\\Footer.js",[],"C:\\Users\\Brilyan Ilham\\Desktop\\foodparadise\\src\\pages\\Home.js",[],{"ruleId":"45","replacedBy":"46"},{"ruleId":"47","replacedBy":"48"},"no-native-reassign",["49"],"no-negated-in-lhs",["50"],"no-global-assign","no-unsafe-negation"]
